# Improve performance of record reader by using Vec::with_capacity